⭐️ Add status display
